improved translations; improved workflow; adapted dev.txt
Workflow: up to now, 'make use' sorted translations.tsv in-place, then
used it to generate content.js. The separation was done so the sorted
file can be used for different browsers.

Now translations.tsv can contain comments and empty lines and will not
be overwritten by 'make use'. That means you can keep related
translations together. For example, the 'generate growth footprint' page
has some strings that only appear on that page, so I didn't want it to
be strewn across the file.